Forgot to commit for a while.
Project pretty much almost finished, just needs styling and text
written out for actions to take.

The program now allows the user to pick an emotion, pick an action (
currently actions are not written out) that will let them deal with that
emotion, and then they will be taken to a final screen that explains
what emotion and action that user picked.